requirements.txt
# Narrative Atlas — Dependencies
# Install: pip install -r requirements.txt

# Core data handling
pandas>=2.0
numpy>=1.24

# Statistics and correlation
scipy>=1.10

# Classical ML pipeline
scikit-learn>=1.3

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Sentiment: lexicon-based
vaderSentiment>=3.3

# Sentiment: transformer-based
transformers>=4.30
torch>=2.0

# Financial data
yfinance>=0.2

# Dataset loading
datasets>=2.14

# Model serialization
joblib>=1.3

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65

# Testing
pytest>=7.4
pytest-cov>=4.1
